+++ src/spatial/index.ts
@@ -1,0 +1,3 @@
+export * from './spatialClient';
+export * from './types';
+export * from './viewport';